Adds support for passing props in Android
Anything passed in a navigator event in a 'passProps' object will be serialized into the Java side Screen object as a HashMap<String, Object>. When RctView starts the React root the props are deserialized into the Bundle passed as props, along with the navigation info. Due to React and Java limitations, JS arrays of arrays are deserialized as dictionaries by index, and arrays of multiple types are not supported.