Reset to (#671)
* Don’t create SideMenu button automatically on Android

* Remove left button and add it dynamically

* Fix resetTo animation on Android

Up until now resetTo would destroy the current stack and create a new
one. This was an easy implementation but resulted in jurky
(none existent) animations.
This commit changes resetTo behavior. Instead of recreating the stack,
the we reset to is pushed to the stack, and the other elements are
removed when show animation ends.