Fixing NPE found on issue #2695 (#2976)
There's a roundabout manner of obtaining the react context on initialisation that is null on some instances (https://github.com/wix/react-native-navigation/issues/2695), including errors on startup that prevents debugging on console. (For instance, throwing an error on index before starting the app)

It is solved by sending the same context given on method onReactContextInitialized, instead of discarding it and calling getReactContext.