Apply resolved options each time a child controller appears
This commit fixes #4045, which revealed a design flaw in the options applying logic.
Child options are applied when a ViewController.onViewAppeared is invoked.
A ViewController's appearance is determined from a GlobalLayoutListener.

Consider the following layout hierarchy:
BottomTabs > Stack > Component

onViewAppeared will be triggered for the BottomTabs first, then for the Stack and last for the component.
This means that if the Stack and the Component options will be applied consecutively which in some cases
might case flickering, mostly when drawBehind options are applied.